more code clean in prepration for finishing Top N queries
